// src/types.d.ts
import { Dispatch } from "react";

export interface BooksContextType {
  books: Library[];
  selectedBook: Library | null;
  loading: boolean;
  getBooks: () => void;
}

export interface Books {
  library: Library[];
}

export interface Library {
  book: Book;
}

export interface Book {
  title: string;
  pages: number;
  genre: string;
  cover: string;
  synopsis: string;
  year: number;
  ISBN: string;
  author: Author;
}

export interface Author {
  name: string;
  otherBooks: string[];
}

export interface Books {
  library: Library[];
}

export interface Library {
  book: Book;
}

export interface Book {
  title: string;
  pages: number;
  genre: string;
  cover: string;
  synopsis: string;
  year: number;
  ISBN: string;
  author: Author;
}

export interface Author {
  name: string;
  otherBooks: string[];
}